move container models to common models
