+++ mex/common/identity/backend_api.py
@@ -1,9 +1,9 @@
 from functools import cache
 
 from mex.common.backend_api.connector import BackendApiConnector
-from mex.common.backend_api.models import ItemsContainer
 from mex.common.identity.base import BaseProvider
 from mex.common.identity.models import Identity
+from mex.common.models import ItemsContainer
 from mex.common.types import Identifier, MergedPrimarySourceIdentifier
